+++ docs/changes.rst
@@ -8,6 +8,7 @@ Change Log
   * added :meth:`.unlink` as synonym for :meth:`.remove`
   * added :meth:`.normalize`
   * added :meth:`.get_r` to recursively copy remote directories to a local path
+  * added :attr:`.pwd` to return the current working directory
 
 
 * 0.2.6 (current, released 2014-05-17)